checkpoint

+++ extern/ldm_zero123/modules/diffusionmodules/model.py
@@ -24,7 +24,9 @@ def get_timestep_embedding(timesteps, embedding_dim):
     emb = math.log(10000) / (half_dim - 1)
     emb = torch.exp(torch.arange(half_dim, dtype=torch.float32) * -emb)
     emb = emb.to(device=timesteps.device)
-    emb = timesteps.float()[:, None] * emb[None, :]
+    emb = timesteps[:, None] * emb[None, :]
+
+    # emb = timesteps.float()[:, None] * emb[None, :]
     emb = torch.cat([torch.sin(emb), torch.cos(emb)], dim=1)
     if embedding_dim % 2 == 1:  # zero pad
         emb = torch.nn.functional.pad(emb, (0, 1, 0, 0))
