+++ extern/ldm_zero123/modules/diffusionmodules/util.py
@@ -186,7 +186,8 @@ def timestep_embedding(timesteps, dim, max_period=10000, repeat_only=False):
             * torch.arange(start=0, end=half, dtype=torch.float32)
             / half
         ).to(device=timesteps.device)
-        args = timesteps[:, None].float() * freqs[None]
+        args = timesteps[:, None] * freqs[None]
+        # args = timesteps[:, None].float() * freqs[None]
         embedding = torch.cat([torch.cos(args), torch.sin(args)], dim=-1)
         if dim % 2:
             embedding = torch.cat(
